# backend-golang/docs/swagger.yaml
basePath: /api
definitions:
...
  domain.Message:
    properties:
      created_at:
        type: string
      id:
        type: string
      payload:
        example: '{"key":"value"}'
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for Sample Stack Golang.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sample Stack Golang API
  version: "1.0"
paths:
  /messages:
    get:
      consumes:
      - application/json
      description: Get messages from all tenants with pagination
      parameters:
      - description: 'Number of messages to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all messages
      tags:
      - messages
...
  /tenants/{tenant_id}/messages:
    get:
      consumes:
      - application/json
      description: Get all messages for a specific tenant with pagination
      parameters:
      - description: Tenant ID
        in: path
        name: tenant_id
        required: true
        type: string
      - description: 'Number of messages to return (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get messages by tenant
      tags:
      - messages
    post:
      consumes:
      - application/json
      description: Create a new message for a specific tenant
      parameters:
      - description: Tenant ID
        in: path
        name: tenant_id
        required: true
        type: string
      - description: Message Information
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/domain.Message'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Message'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new message
      tags:
      - messages
  /tenants/{tenant_id}/messages/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a message for a tenant
      parameters:
      - description: Tenant ID
        in: path
        name: tenant_id
        required: true
        type: string
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a message
      tags:
      - messages
    get:
      consumes:
      - application/json
      description: Get a specific message by its ID for a tenant
      parameters:
      - description: Tenant ID
        in: path
        name: tenant_id
        required: true
        type: string
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Message'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get message by ID
      tags:
      - messages
    put:
      consumes:
      - application/json
      description: Update an existing message for a tenant
      parameters:
      - description: Tenant ID
        in: path
        name: tenant_id
        required: true
        type: string
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Message Information
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/domain.Message'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Message'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a message
      tags:
      - messages
...
schemes:
- http
- https
swagger: "2.0"